+++ src/responses/raw/update.rs
@@ -2,7 +2,7 @@ use responses::raw::message::Message;
 
 #[derive(Deserialize, Debug)]
 pub struct Update {
-    pub update_id: i32,
+    pub update_id: i64,
     pub message: Option<Message>,
     pub edited_message: Option<Message>,
     pub channel_post: Option<Message>,
